requirements.txt
discord.py
aiosqlite
python-dotenv
aiohttp
qrcode